fix: Remove filter data model
